+++ theodb_rs/src/lib.rs
@@ -38,6 +38,7 @@ mod nl;
 mod pg;
 mod pq;
 mod vec;
+mod vectorizer;
 
 mod api;